# cimrms-api/openapi.yaml
openapi: 3.0.3
info:
  title: MRDI CIM-RMS API
  description: CIM 需求管理系统（RMS）— Normal Change 需求工作流 + SPC 统计过程控制（Sprint 4）
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Local dev
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pagination:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
        totalPages: { type: integer }

    # ── Auth ─────────────────────────────────────────────────────────────
    DevLoginQuery:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        role: { type: string, enum: [admin, auditor, editor, viewer], default: editor }
        name: { type: string, default: 'Dev User' }
        department: { type: string, default: CIM }
    LoginResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/CurrentUser' }
    CurrentUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [admin, auditor, editor, viewer] }
        department: { type: string }
        teams: { type: array, items: { type: string } }

    # ── Request ───────────────────────────────────────────────────────────
    RequestStatus:
      type: string
      enum:
        - pending_manager
        - manager_rejected
        - pool
        - scheduled
        - in_development
        - pending_uat
        - pending_deploy
        - pending_acceptance
        - closed
        - cancelled
    RequestUrgency:
      type: string
      enum: [P1, P2, P3]
    RequestType:
      type: string
      enum: [feature, config, report, integration, other]
    Request:
      type: object
      properties:
        id: { type: string, format: uuid }
        requestNo: { type: string }
        title: { type: string }
        type: { $ref: '#/components/schemas/RequestType' }
        urgency: { $ref: '#/components/schemas/RequestUrgency' }
        status: { $ref: '#/components/schemas/RequestStatus' }
        relatedSystem: { type: string, nullable: true }
        description: { type: string }
        attachmentIds: { type: array, items: { type: string } }
        relatedIncidentId: { type: string, nullable: true }
        expectedCompletion: { type: string, format: date-time, nullable: true }
        submitterEmail: { type: string }
        submitterDept: { type: string }
        team: { type: string, nullable: true }
        estimatedDeployAt: { type: string, format: date-time, nullable: true }
        assigneeEmail: { type: string, nullable: true }
        slaPercent: { type: number, description: 'SLA 消耗百分比，超过 100 = 超时' }
        isOverdue: { type: boolean }
        statusEnteredAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
    RequestPage:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            items: { type: array, items: { $ref: '#/components/schemas/Request' } }
...
    CreateRequestBody:
      type: object
      required: [title, type, urgency, description]
      properties:
        title: { type: string, minLength: 2, maxLength: 200 }
        type: { $ref: '#/components/schemas/RequestType' }
        urgency: { $ref: '#/components/schemas/RequestUrgency' }
        relatedSystem: { type: string, maxLength: 50 }
        description: { type: string, minLength: 10, maxLength: 5000 }
        attachmentIds: { type: array, items: { type: string, format: uuid }, default: [] }
        relatedIncidentId: { type: string, maxLength: 32 }
        expectedCompletion: { type: string, format: date-time }
    CreateRequestResponse:
      type: object
      properties:
        id: { type: string }
        requestNo: { type: string }
        status: { type: string }
        submitterEmail: { type: string }
        createdAt: { type: string, format: date-time }
...
    CommentBody:
      type: object
      required: [comment]
      properties:
        comment: { type: string, minLength: 1, maxLength: 2000 }
        attachmentIds: { type: array, items: { type: string, format: uuid }, default: [] }
...
paths:
...
  /v1/auth/dev/login:
    get:
      tags: [Auth]
      summary: Dev login (GET, development only)
      parameters:
        - in: query
          name: email
          required: true
          schema:
            type: string
            format: email
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, auditor, editor, viewer]
            default: editor
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: department
          schema:
            type: string
      responses:
        '200':
          description: Dev login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
    post:
      tags: [Auth]
      summary: Dev login (POST, SPA compatible)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginQuery'
      responses:
        '200':
          description: Dev login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Login alias (POST)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DevLoginQuery'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
...
  /v1/requests:
    get:
      tags: [Requests]
      summary: List requests with filters
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: urgency
          schema:
            type: string
            enum: [P1, P2, P3]
        - in: query
          name: type
          schema:
            type: string
            enum: [feature, config, report, integration, other]
        - in: query
          name: submitter
          schema:
            type: string
        - in: query
          name: assignee
          schema:
            type: string
        - in: query
          name: team
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: view
          schema:
            type: string
            enum: [all, mine, pending_approval, pending_action, submitted_by_me]
            default: all
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Request list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestPage'
    post:
      tags: [Requests]
      summary: Create a new request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestBody'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRequestResponse'
        '400':
          description: Validation error
  /v1/requests/{id}:
    get:
      tags: [Requests]
      summary: Get request detail with availableActions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
        '404':
          description: Not found
...
  /v1/requests/{id}/comment:
    post:
      tags: [Requests]
      summary: Add comment (viewer+)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentBody'
      responses:
        '200':
          description: Comment added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  ok:
                    type: boolean
...